backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pytesseract==0.3.10
Pillow==10.2.0
